requirements.txt
asgi-redis==0.10.0
asgiref==0.11.2
autobahn==0.13.1
channels==0.11.0
daphne==0.10.3
Django==1.9.5
django-crudbuilder==0.1.11
django-extensions==1.6.1
django-tables2==1.1.6
msgpack-python==0.4.7
oauthlib==1.0.3
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
Twisted==16.1.1
txaio==2.3.1
wheel==0.26.0
zope.interface==4.1.3
